+++ Jumpman/Main.h
@@ -9,12 +9,12 @@ extern "C" {
 #endif
 
 bool IsGameFrozen(void);
-long GetCurrentFps(void);
+int32_t GetCurrentFps(void);
 double GetCurrentMaxUpdateFrameTime(void);
 double GetCurrentMaxDrawAndSwapFrameTime(void);
-long GetLastKeyPressed(void);
-int GetKeyBinding(size_t binding_index);
-void SetKeyBinding(size_t binding_index, int value);
+int32_t GetLastKeyPressed(void);
+int32_t GetKeyBinding(size_t binding_index);
+void SetKeyBinding(size_t binding_index, int32_t value);
 void SaveSettings(void);
 
 #if defined(__cplusplus)
